refactor: move message_type field outside of payload
This moves the message_type field outside of the payload object and into
the base message object. It allows use of Generic[PayloadT] and better
type declarations for functions that accept messages.

messages and payloads were both moved out of the models.py and into
their own messages.py class.